Times are now all presented to users in the local time, but processed in UTC.
This accepts time from the server in UTC, with an additional offset. All diffs are done based on UTC.

The other additional issue this deals with is the browser's local timezoneOffset in the date object. This is handled on parsing the date and then on sending it to the server again.

Note: this probably breaks saving dates - this will be handled in a latter commit.