Adjust start/end times by the timezone offset when creating new events.
This is already being done when events are edited, but it is hard to
do this for new events, because they can happen at a new place, at
which we do not know the timezone on the client. Therefore we need to
handle this case here.

An option for the future, might be to actually know what the timezone
of the new place is on the client side, then both edit and save can
send the same time. For now though, it is not quite identical.