Get all participants in matching events.
Previously, we only returned participants that had sufficient importance.
Now we return all participants regardless of their importance, so long
as one participant does have enough importance to make the event they are
in captured.

This requires doing an inner query to get the relevant events and an outer
query that then fetches all the participants for the event.