when an event changes update make sure it updates everywhere
This includes the map marker and the thing summary, and required some
data to be sent in a better format, namely PUT /event now returns the
event id and search returns event ids as numbers, not strings.

This has the side effect of causing redraws more often. Monitor this and
see if it ever becomes an issue. An enhancement might be to only update
if there was been a more substantial change.