Do not lose the /state path on page load
On a new page load, the highlight will be taken from the URL, and can
load the thing_summary before the search query has come back. Therefore
only get into updateHighlight() which will set the index of the
thing_summary and hence the index if it is ready for it.